+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next release
 * Removed support for Python 2
+* CSS for Column layout object in Bootstrap 4 template pack changed to 'col-md'. Default is now over ridden when another 'col' class is added to css_class.
 
 ## 1.8.1 (2019-11-22)